Add support for batched analysis using Anthropic on AWS Bedrock
This is much, much faster than sequential analysis. On a codebase
of approx 12,500 files I estimate that sequential analysis will
take ~10 days. In batch mode it takes less than a day.

However, the quality of the output is not as good so I need to
investigate the root cause.

I have hardcoded quite a lot of Anthropic/Bedrock-specific types and
logic into the batch services, but it should be easy enough to move
that out into a BedrockAnthropic batch_client and keep the main
codebase generic.